CNN RNN split
