--- .gitignore
@@ -1,2 +1,0 @@
-Cifar10/
-__pycache__/
