+++ CNN/.gitignore
@@ -1,0 +1,3 @@
+Cifar10/
+__pycache__/
+saved_models/alex.pt
