+++ CNN/ensamble_data/predictions_max_model_series_aug
