+++ CNN/ensamble_data/predictions_mean_model_series_aug
