+++ CNN/ensamble_data/predictions_model1_model_series
