+++ CNN/ensamble_data/predictions_model2_model_series
