+++ CNN/ensamble_data/predictions_model3_model_series
